feat: 실시간 문제집 API 연동 테스트
