+++ src/app.feature/quiz/api/index.ts
@@ -1,0 +1,8 @@
+import client from '@app.modules/api/client';
+import { IUpdatedQuiz } from '../types';
+
+export const temp = '';
+
+export const fetchGetUpdatedQuizzes = (): Promise<IUpdatedQuiz[]> => {
+	return client.get('/question-book/list/live').then((response) => response.data.result);
+};
