+++ src/app.feature/quiz/component/card/UpdatedQuizCard.tsx
@@ -2,16 +2,19 @@ import React from 'react';
 import PersonIcon from '@assets/quiz/person.svg';
 import ArrowRightIcon from '@assets/quiz/arrow-right.svg';
 import { transformSecond } from '@app.modules/util';
+import { IUpdatedQuiz } from '@app.feature/quiz/types';
 
 interface Props {
 	className?: string;
-	updateTime: number;
-	userNickName: string;
-	studyGroupName: string;
-	goalContent: string;
+	quizData: IUpdatedQuiz;
 }
 
-export default function UpdatedQuizCard({ className, updateTime, userNickName, studyGroupName, goalContent }: Props) {
+export default function UpdatedQuizCard({ className, quizData }: Props) {
+	const nowDate = new Date();
+
+	const createdDate = Date.parse(quizData.questionCreatedAt);
+	const updateTime = (nowDate.getTime() - createdDate) / 1000;
+
 	return (
 		<div className={`px-[18px] bg-background-white rounded drop-shadow-white ${className}`}>
 			<div className="flex justify-end pt-[10px]">
@@ -25,11 +28,11 @@ export default function UpdatedQuizCard({ className, updateTime, userNickName, s
 					</div>
 					<div>
 						<div className="flex">
-							<span className="text-small text-brown font-regular">{userNickName}</span>
+							<span className="text-small text-brown font-regular">{quizData.userNickName}</span>
 							<hr className="border border-brown h-[9px] my-auto ml-[6px] mr-[9px]" />
-							<span className="text-small text-brown font-bold">{studyGroupName}</span>
+							<span className="text-small text-brown font-bold">{quizData.studyGroupName}</span>
 						</div>
-						<p className="text-body1 font-medium">{goalContent}</p>
+						<p className="text-body1 font-medium">{quizData.goalContent}</p>
 					</div>
 				</div>
 				<div className="my-auto">
