+++ src/app.feature/quiz/types/index.ts
@@ -1,0 +1,9 @@
+export interface IUpdatedQuiz {
+	questionBookId: number;
+	goalContent: string;
+	userNickName: string;
+	userProfileImageUrl?: string | null;
+	studyGroupName: string;
+	questionCreatedAt: string;
+	solved: boolean;
+}
